refactor: extract readCoordinator and autoReadPolicy from readPdf handler
readPdf.ts: 955 → 238 lines (-75%). Fat controller split into:
- autoReadPolicy.ts (214 lines): auto-read decision logic
- readCoordinator.ts (529 lines): pipeline orchestration
- readPdf.ts (238 lines): thin transport handler

All 348 tests pass. No behavior changes.
